requirements.txt
PyMySQL==1.0.2
Flask-Migrate==3.1.0
SQLAlchemy==1.4.36
Flask-JWT-Extended==4.4.0
APScheduler==3.9.1
asn1tools==0.163.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
bitstruct==8.14.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
click-default-group==1.2.2
Deprecated==1.2.13
diskcache==5.4.0
dnspython==2.2.1
Flask==2.1.2
Flask-APScheduler==1.12.3
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
ifaddr==0.1.7
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
kazoo==2.5.0
MarkupSafe==2.1.1
nacos-sdk-python==0.1.7
packaging==21.3
profiling==0.1.3
prompt-toolkit==3.0.29
py-eureka-client==0.10.5
pykafka==2.8.0
pymongo==4.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-redis-lock==3.7.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
redis==4.3.1
requests==2.27.1
retrying==1.3.3
six==1.16.0
starkbank-ecdsa==1.1.1
tabulate==0.8.10
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
urwid==2.1.2
valuedispatch==0.0.1
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0